Improve polyfilling
Importing @babel/polyfill (deprecated as of Babel 7.4) in our entry
files imports all polyfills, even if they aren't being used, and
pollutes the global namespace.

This commit uses the runtime-corejs3 as the runtime used during the
transform-runtime plugin, which will both only polyfill what is used and
not alter the global namespace.