Run proof-of-concept on builds of userbase-js and userbase-server
Currently proof-of-concept is compiling the source code of userbase-js
during the webpacking of the client.

This is happening because the userbase-js dependency is linked (either
by npm-link or using "file:" in package.json) to the local filesystem,
causing `exclude: '/node_modules/'` rule to not apply (since the path
resolves to the one on the local file system).

This was leading to subtle problems where modifications to how
userbase-js was compiled (in it's own package) were breaking
proof-of-concept, again because proof-of-concept was transpiling the
userbase-js source code itself using it's webpack configuration.

This PR uses changes the proof-of-concept webpack configuration to
properly exclude the userbase-js package when linked to the local file
system. This means that the dependencies need to point to compiled
versions of userbase-js.

In order to do that I linking, I'm also proposing using npm-link:
```
npm link ../../npm_build/userbase-js
npm link ../../npm_build/userbase-server
```

Rather than linking using relative paths:
```
  "userbase-js": "file:../../npm_build/userbase-js",
  "userbase-server": "file:../../npm_build/userbase-server"
```

This makes makes it so that package stands on its own and the option to
use a local build is up to the developer.

A couple things to consider now that proof-of-concept no longer compiles
the userbase-js source:
1) npm start will no longer work unless userbase-js is built
2) changes to userbase-js will need to be rebuilt to picked up

These considers can both be cleaned up by (optionally) watching both
userbase-js projects as part of start.