Add a new paramter allowPromotionCodes to purchaseSubscription and updatePaymentMethod
This commit adds another property under this "Advanced" section called "Allow promotion codes in Checkout" with a link to the documentation. This value `allowPromotionCodes` is stored on the app in the apps table. It is passed to stripe during `redirectToCheckout` for `purchaseSubscription` and `updatePaymentMethod`. It allows users to provide promotion codes.